Fixed UI not updating when database updates
