+++ app/src/main/java/com/example/uitnotify/Worker.kt
@@ -1,6 +1,7 @@
 package com.example.uitnotify
 
 import android.content.Context
+import android.content.SharedPreferences
 import android.util.Log
 import androidx.compose.ui.platform.LocalContext
 import androidx.work.Worker
@@ -14,11 +15,26 @@ import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
 import java.io.IOException
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+import java.time.temporal.ChronoUnit
 
-class ArticleWorker(appContext: Context,
-                    workerParams: WorkerParameters) : Worker(appContext, workerParams) {
+class ArticleWorker(appContext: Context, workerParams: WorkerParameters) : Worker(appContext, workerParams) {
+
+    private val articleDao = AppDatabase.getDatabase(applicationContext).articleDao()
+    private val sharedPreferences: SharedPreferences = appContext
+        .getSharedPreferences("ArticleWorkerPrefs", Context.MODE_PRIVATE)
+    private val dateTimeFormatter: DateTimeFormatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME
+
     override fun doWork(): Result {
         Log.d("ArticleWorker", "doWork() started")
+        val lastDownloadTime = getLastDownloadTime()
+        val now = LocalDateTime.now()
+        val timeSinceLastDownload = ChronoUnit.MINUTES.between(lastDownloadTime, now)
+        if (timeSinceLastDownload < 1) {
+            Log.d("ArticleWorker", "doWork: Not enough time has passed since last download. Skipping.")
+            return Result.success()
+        }
         var result: Result = Result.failure()
         try {
             runBlocking {
@@ -27,14 +43,14 @@ class ArticleWorker(appContext: Context,
                     Log.d("ArticleWorker", "coroutineScope started")
                     launch {
                         Log.d("ArticleWorker", "launch started")
-                        val articleDao = AppDatabase.getDatabase(applicationContext).articleDao()
-                        articleDao.deleteAllArticles()
-                        Log.d("ArticleWorker", "deleteAllArticles() completed")
+                        // articleDao.deleteAllArticles()
+                        // Log.d("ArticleWorker", "deleteAllArticles() completed")
                         val articles = downloadArticles()
                         Log.d("ArticleWorker", "downloadArticle() completed, articles size: ${articles.size}")
                         sendNotification(applicationContext, "UIT Notify", "Downloaded ${articles.size} articles")
                         saveArticlesToDatabase(articles)
                         Log.d("ArticleWorker", "saveArticlesToDatabase() completed")
+                        saveLastDownloadTime(now)
                         result = Result.success()
                         Log.d("ArticleWorker", "Result.success()")
                     }
@@ -55,6 +71,7 @@ class ArticleWorker(appContext: Context,
     }
 
     private suspend fun downloadArticles(): List<Article> = withContext(Dispatchers.IO) {
+        Log.d("Schedule", "Scheduling periodic article download")
         Log.d("ArticleWorker", "downloadArticles() started")
         val url = "https://student.uit.edu.vn/thong-bao-chung"
         val articles = mutableListOf<Article>()
@@ -95,6 +112,20 @@ class ArticleWorker(appContext: Context,
             throw e
         }
         Log.d("ArticleWorker", "saveArticlesToDatabase() completed")
+    }
+
+    private fun getLastDownloadTime(): LocalDateTime {
+        val lastDownloadTimeString = sharedPreferences.getString("lastDownloadTime", null)
+        return if (lastDownloadTimeString != null) {
+            LocalDateTime.parse(lastDownloadTimeString, dateTimeFormatter)
+        } else {
+            LocalDateTime.now().minusYears(1)
+        }
+    }
+
+    private fun saveLastDownloadTime(time: LocalDateTime) {
+        val timeString = time.format(dateTimeFormatter)
+        sharedPreferences.edit().putString("lastDownloadTime", timeString).apply()
     }
 }
